requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
fastapi==0.75.1
h11==0.13.0
httptools==0.4.0
idna==3.3
Naked==0.1.31
pycryptodome==3.14.1
pydantic==1.9.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
